added Sked class, developers.txt
each of the spacecraft components has been added with a power variable indicating if that component is powered ON or OFF (true or false)

sensor.power
controller.power
actuator.power

the update() methods of each components are also updated
to include power as a condition, when power is not TRUE then data variable of the component is not updated

added new class Sked()
Sked refers to the command line based schedule of the AOCS subsystem, the sked file is initialized by a sked.txt file

added new methods to Spacecraft() class to make use of Sked()
loadSked() -- load sked file into spacecraft and initialize required parameters
updateUnixTime() -- built-in AOCS OBC process cycle ticker that ticks every cycle (which value is specified, default = 1 sec) every cycle runs processSked() and updateComponents() methods
processSked() -- processes the sked command lines, based on the parameters intialized, this interprets the commands which uses the COMMANDEXEC() method
COMMANDEXEC() -- interprets the command and argument for each command line given
updateComponents() -- calls all update methods: updateSensors(), updateControllers(), updateActuators()

in the LEOSS() class, the update methods under advance1timestep() method has been replaced with the updateUnixTime() method. This shows that for each cycle of the OBC the components are read/run (sensor, controllers, actuators) for the spacecraft.

TO-DO: add a test to verify the results of adding the Sked class
might need to add a feature in which if no Sked file is available then it will create a new one as default. also need to add a Log file to record the commands that are processed by the OBC during the run.

added developers.txt -- in case i forget how to run the tests and update the dependencies of the project